Versión buena

+++ src/Controller/PostController.php
@@ -7,9 +7,9 @@ use App\Entity\Post;
 use App\Form\PostType;
 use DateTime;
 use Doctrine\ORM\EntityManagerInterface;
-use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
+use Symfony\Component\Finder\Exception\AccessDeniedException;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -96,10 +96,37 @@ class PostController extends AbstractController
     #[Route('/edit/{id}', name: 'app_posts_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Post $post, EntityManagerInterface $entityManager): Response
     {
+        // Verificar si el usuario autenticado es el creador del post
+        if ($this->getUser() !== $post->getIdCreador()) {
+            throw new AccessDeniedException('You are not allowed to edit this post.');
+        }
+
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+
+            // Manejar la carga de la imagen
+            $imagen = $form['imagen']->getData();
+
+            if ($imagen) {
+                // Generar un nombre único para la imagen
+                $nombreArchivo = md5(uniqid()) . '.' . $imagen->guessExtension();
+
+                // Mover la imagen al directorio de destino
+                try {
+                    $imagen->move(
+                        $this->params->get('images_directory_posts'),
+                        $nombreArchivo
+                    );
+                } catch (FileException $e) {
+                    // Manejar errores si la carga de la imagen falla
+                }
+
+                // Establecer el nombre de la imagen en el post
+                $post->setImagen($nombreArchivo);
+            }
+            
             $entityManager->flush();
 
             return $this->redirectToRoute('app_posts', [], Response::HTTP_SEE_OTHER);
@@ -114,6 +141,11 @@ class PostController extends AbstractController
     #[Route('/{id}', name: 'app_posts_delete', methods: ['POST'])]
     public function delete(Request $request, Post $post, EntityManagerInterface $entityManager): Response
     {
+        // Verificar si el usuario autenticado es el creador del post
+        if ($this->getUser() !== $post->getIdCreador()) {
+            throw new AccessDeniedException('You are not allowed to delete this post.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$post->getId(), $request->request->get('_token'))) {
             $entityManager->remove($post);
             $entityManager->flush();
